Fix QR code popup not showing after being dismissed by the user
Happened when a non-payment QR code was detected and the user
dismissed the popup. Afterwards no more popups were shown because
the `currentQRCodePopup` was not cleared and calling it's `hide`
method did not invoke the completion callback because it was already
removed from the superview.

Fixed it by calling the `didDismiss` block also when the user
dismisses the popup.